+++ internal/services.go
@@ -293,7 +293,7 @@ type buildingQueueItem struct {
 	cityID         string
 	queuedEpoch    int64
 	durationSec    int64
-	targetLevel    int32
+	targetLevel    int64
 	targetBuilding string
 }
